Finished! (hopefully, needs testing)

+++ src/main/java/app/DAOs/IDAO.java
@@ -5,7 +5,7 @@ import java.util.Set;
 public interface IDAO <T>{
     T create(T t);
     Set<T> get();
-    T getByID(Long id);
+    T getByID(Integer id);
     T update(T t);
     Integer delete(T t);
 }
